Add support for new subgraph IDs (#87)
A upgrade to the GNS contracts will change the format of subgraph IDs to
be base58-encoded keccak hashes of the accoount ID and sequence ID that
make up the current subgraph IDs. To support the new IDs, and maintain
backward compatibility, All subgraph IDs will be parsed and stored in
the new ID format.