Use fresh context for each indexer (#97)
Say we have some query that has no block requirement, so making the query deterministic will set a block hash based on the latest block for that indexer. If we need to try another indexer, then the context will be carried over and the next indexer may be queried with a block that is greater than its indexing status. This will result in unnecessary retries and penalties for many indexers that happen to be selected after the first for any query.

See also: https://github.com/graphprotocol/agora/pull/31