Simplify query logs
Previously, the logs used for data science and production investigations
were spread across "client query result", "indexer query result", and
the `indexer_response_err` fields. This collects all this information
throughout the lifecycle of a query. When a response is returned to the
client, there will now be a singe "Client query result" event, and a set
of "Indexer attempt" events. Logs for ISA scoring remain separate.