Adjust log levels (#165)
Trace logging is enabled in production, since the ISA socring logs are
used for data science. However log storage is expensive, and slows down
monitoring jobs as well. So this raises the level of scoring logs to
debug and drops some others down to trace, with the intention of
disabling trace logs on mainnet & testnet gateways.