remove unnecessary abstraction between SubgraphDeployments and its inputs (#266)
- Unfortunately this is a situation where code speaks faster than words.
Mostly, the structure & inputs of `SubgraphDeployments` is unnecessarily
obscured. So I just removed that and did the necessary tidying up.
- On the topic of lifetimes for parsing through the response. We can
just walk a reference to the response in `parse_current_deployments` to
build up the map, since copying the data we're trying to extract is
fairly cheap (just 32-byte copies everywhere). Then on the second
function we can move the response and therefore move the data we want to
extract, since we might as well avoid the copy at this point.