feat(#267): create subscriptions subgraph client (#272)
# Description

Closes #267 

Create a subgraph client that can query the active subscriptions and use
the data to map the active subscription to a user. This will allow us
to:
- receive a request ticket (not part of this pr)
- parse/validate (not part of this pr)
- check the stored active subscription map by the user address parsed
from the request ticket (in this pr)
- validate the subscription is active (in this pr)

## Notes/looking for feedback

- I added a `subscriptions_subgraph` param to `opt` to populate the url,
similar to how we do `network_subgraph`
- I see this value set in `local-network-environments`, but do I need to
set it anywhere else
- the client is not wired up yet
- need to update `local-network-environments` (will update next) to
build up subgraph/contracts and add url in gateway
- I added a helper fn that checks if the user has an active
subscription, and if the current timestamp (passed in) is in the valid
range for the active subscription. This might be overkill, but thought
it may come in handy.
- I created a `subgraph_client` and moved the query logic into it so I
didn't have to replicate between the `network_subgraph` and
`subscriptions_subgraph`
- I updated the `network_subgraph.poll_network_params` to return a
single `GraphNetwork` instance. There is only 1 graph network instance
on the subgraph, and its id is "1", we do this same thing in
explorer/studio. Small change.