Overhaul Reporting for Data Science (#295)
I have made multiple attempts in the past to clean this up, which all
failed to meaningfully improve the situation. The primary goal of this
is to track information about queries as it becomes available in the
code path, and moving the cumbersome backwards compatibility concerns
somewhere else. This approach adds a custom layer to the `tracing`
framework that collects events from special targets to eventually
assemble the logs and Kafka messages for the data science team. It's a
bit rough now. But I think it's finally a good starting point to
facilitating the changing needs for standardization around gateway
reporting.

This also adds subscription query Kafka messages.