Work around compiler bug
This tweaks the Agora context type we use in order to avoid a recurring
Rust compiler bug (https://github.com/rust-lang/rust/issues/71723) which
makes it suddenly impossible to carry the context across an await. This
comes at the cost of increasing String allocations. Though the use of
snmalloc mitigates this cost somewhat.