Resolve subgraph versions from IPFS (#323)
This adds subgraph versions labels to deployments, which will eventually
be used to support more graceful subgraph upgrades. We could have used
the [SubgraphVersion
label](https://github.com/graphprotocol/graph-network-subgraph/blob/master/schema.graphql#L466)
directly from the subgraph, but that would create a dependency on the
version of the network subgraph that uses IPFS. We are avoiding that and
parsing from IPFS manually so that the non-IPFS version of the network
subgraph can be put on the network for gateways to use.