feat(graph-gateway): added indexers status public indexing statuses client (#375)
This PR aims to improve the modularity and testability of the current
gateway codebase. With this, I believe we'll reduce the probability of
git conflicts, as modules are smaller, and improve the overall quality
of the code as maintainability and correctness.

Following the structure/pattern established in #372, I implemented
another query under the `indexers_status` module that the gateway
performs to the `/status` endpoint: `indexingStatuses`.

- [x] Added the query and response serialization/deserialization logic.
- [x] Added a client method.
- [x] Added unit and integration tests.

A follow-up PR will be where this "client code" replaces the existing
code in the `indexing.rs` file.