refactor: UDecimal18 (#417)
Converting between monetary types is surprisingly tricky business. This
makes a bunch of changes I've wanted for quite some time.
- Rewrite the decimals.rs to use the `alloy_primitives::U256`. This
removes a lot of code that is no longer necessary to compensate for
issues with `primitive_types::U256`.
- Restrict the decimals to 18. The old design was a bit too cute since
it resulted in some conversions needing to be fallible even though that
was unnecessary for our use-case.
- Make `GRT` and `USD` separate types rather than aliases for the same
type. It was a dangerous decision to allow these to be used
interchangeably. exchange_rate.rs was a lingering example of some type
confusion (luckily benign), which is now resolved.