chore(deps): bump toolshed version to v0.5.0 (#478)
A new release for the `toolshed` utility crate is out. This PR pins the
`toolshed` crate version to the `v0.5.0`.

> [!WARNING]
> Using the latest version (or the `main` branch), probably unstable, is
discouraged as it introduces an implicit, uncontrolled, and nontraceable
dependency change.
>
> It is preferable to use a git tag (or git revision hash) so the
version upgrade is traced as part of the development process in the git
history, allowing to revert the dependency changes as a quick solution
if the upgrade introduced a regression.