refactor: migrate to new ISA implementation (#588)
This switches to a rewrite of indexer-selection that has been stripped
down to only concerning itself with picking candidates out of a valid
set. The new implementation removes a lot of long-standing
inefficiencies, and will be easier to iterate on changes to the ISA.

See https://github.com/edgeandnode/candidate-selection

A bunch of integration testing and production performance comparisons
will happen before this is released.