allow startup when some chain RPCs are failing (#615)
Based on recent incidents, it is generally more important to allow
restarts to proceed than it is to make sure all block caches are ready.
Especially now that there's a 2 second interval for making chain head
requests. This also reduces the error level to WARN when the block
resolution request may fail due to bad user input.