fix: inefficiencies in processing Agora contexts (#747)
This changes the Agora context to only rely on a `&str` into the client
request document to avoid unnecessary `String` allocations, and removes
the need to copy the context for every indexer selected. Most of the
changes are dedicated to indexer request rewrites, to avoid the need for
`Context<_, String>`, while relying on `graphql_parser` to serialize
everything below the top-level selection set fields.