refactor(graph-gateway): initialize receipt signer pools just in time (#807)
As described in #790:

> The current implementation relies on an eventual initialization and
update of the receipt signer's internal mapping between
indexer-deployment pairs, a.k.a. indexing ID, and the largest allocation
associated with this indexing.
> 
> This dependency complicates the request-handling pipeline and can be
simplified by initializing the receipt pools just in time.

This is the first step to changing internal behavior. Once #743 is
merged, I will solve the TODOs introduced in this PR.

Signed-off-by: Lorenzo Delgado <lorenzo@edgeandnode.com>