fix: select deployment outside of indexer-selection (#857)
This should improve user experience by selecting the subgraph version
outside of indexer-selection. Using indexer-selection hasn't been a
perfect fit because it considers factors, such as expected latency, that
do not line up with expectations for version selection. The downside of
this change is that it opens up an abuse vector for indexers that claim
a false indexing status. But we have decided it is reasonable to block
such indexers.